refactor: 상수들에게 as const 키워드 붙이기 (#381)
